+++ test/openBCIBoard-test.js
@@ -245,34 +245,18 @@ describe('openbci-sdk', function () {
       expect(ourBoard2.options.simulatorSerialPortFailure).to.be.true;
     });
     it('should be able to enter sync mode', function () {
-      var ourBoard1, ourBoard2;
-
-      ourBoard1 = new openBCIBoard.OpenBCIBoard({
+      var ourBoard = new openBCIBoard.OpenBCIBoard({
         sntpTimeSync: true
       });
-      expect(ourBoard1.options.sntpTimeSync).to.be.true;
+      expect(ourBoard.options.sntpTimeSync).to.be.true;
 
-      // Verify multi case support
-      ourBoard2 = new openBCIBoard.OpenBCIBoard({
-        sntptimesync: true
-      });
-      expect(ourBoard2.options.sntpTimeSync).to.be.true;
-
-      return Promise.all([
-        new Promise((resolve, reject) => {
-          ourBoard1.once('sntpTimeLock', resolve);
-          ourBoard1.once('error', reject);
-        }),
-        new Promise((resolve, reject) => {
-          ourBoard2.once('sntpTimeLock', resolve);
-          ourBoard2.once('error', reject);
-        })
-      ]).then(() => {
-        ourBoard1.sntpStop();
-        ourBoard2.sntpStop();
+      return new Promise((resolve, reject) => {
+        ourBoard.once('sntpTimeLock', resolve);
+        ourBoard.once('error', reject);
+      }).then(() => {
+        ourBoard.sntpStop();
       }, err => {
-        ourBoard1.sntpStop();
-        ourBoard2.sntpStop();
+        ourBoard.sntpStop();
         return Promise.reject(err);
       });
     });
@@ -525,6 +509,9 @@ describe('openbci-sdk', function () {
       if (spy) spy.reset();
     });
     describe('#connect/disconnect/streamStart/streamStop', function () {
+      it('rejects if already disconnected', function () {
+        return ourBoard.disconnect().should.be.rejected;
+      });
       it('rejects if already connected', function (done) {
         ourBoard.connect(masterPortName).catch(err => done(err));
 
@@ -614,7 +601,7 @@ describe('openbci-sdk', function () {
       });
       afterEach(function (done) {
         if (ourBoard.isConnected()) {
-          ourBoard.disconnect().then(done, () => done());
+          ourBoard.disconnect().then(done, done);
         } else {
           done();
         }
@@ -640,12 +627,10 @@ describe('openbci-sdk', function () {
         var writeSpy1 = sinon.spy(ourBoard.serial, 'write');
         var byteToWrite = k.OBCISDLogStop;
         var writeWhileConnected = function () {
-          var commands = [];
-          while (commands.length < 4) commands.push(byteToWrite);
-          ourBoard.write(commands).then(() => {
+          ourBoard.write(byteToWrite).then(() => {
             if (ourBoard.isConnected()) {
               writeSpy1.reset();
-              setTimeout(writeWhileConnected, 10 * (commands.length - 1));
+              writeWhileConnected();
             } else {
               done('wrote when not connected');
             }
@@ -653,15 +638,17 @@ describe('openbci-sdk', function () {
             if (ourBoard.isConnected()) {
               done(err);
             } else {
-              ourBoard.connect(masterPortName).catch(done);
-              var writeSpy2 = sinon.spy(ourBoard.serial, 'write');
-              ourBoard.once('ready', () => {
-                writeSpy2.should.equal(ourBoard.serial.write);
-                writeSpy1.should.have.not.been.called;
-                writeSpy2.should.have.not.been.calledWith(byteToWrite);
-                writeSpy1.restore();
-                writeSpy2.restore();
-                done();
+              process.nextTick(() => {
+                ourBoard.connect(masterPortName).catch(done);
+                var writeSpy2 = sinon.spy(ourBoard.serial, 'write');
+                ourBoard.once('ready', () => {
+                  writeSpy2.should.equal(ourBoard.serial.write);
+                  writeSpy1.should.have.not.been.called;
+                  writeSpy2.should.have.not.been.calledWith(byteToWrite);
+                  writeSpy1.restore();
+                  writeSpy2.restore();
+                  done();
+                });
               });
             }
           });
@@ -669,13 +656,17 @@ describe('openbci-sdk', function () {
         writeWhileConnected();
         ourBoard.disconnect().catch(done);
       });
-      it('disconnects immediately without performing buffered writes', function (done) {
+      it('disconnects immediately, rejecting all buffered writes', function () {
         var writeSpy = sinon.spy(ourBoard.serial, 'write');
-        ourBoard.write(k.OBCISDLogStop);
-        ourBoard.disconnect().then(() => {
+        return Promise.all([
+          ourBoard.write(k.OBCISDLogStop).should.have.been.rejected,
+          ourBoard.write(k.OBCISDLogStop).should.have.been.rejected,
+          ourBoard.write(k.OBCISDLogStop).should.have.been.rejected,
+          ourBoard.write(k.OBCISDLogStop).should.have.been.rejected,
+          ourBoard.disconnect()
+        ]).then(() => {+
           writeSpy.should.have.not.been.called;
           writeSpy.restore();
-          done();
         });
       });
     });